+++ forui/lib/src/widgets/tile/tile_content.dart
@@ -29,20 +29,21 @@ class FTileContent extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final tile = FTileData.maybeOf(context)!;
-    final FTileData(style: tileStyle, :enabled, :hovered) = tile;
+    final FTileData(style: tileStyle, :enabled, :hovered, :focused) = tile;
 
     final group = extractTileGroup(FTileGroupData.maybeOf(context));
 
-    final FTileStyle(:contentStyle, :dividerStyle) = tileStyle;
+    final FTileStyle(:contentStyle, :dividerStyle, :focusedDividerStyle) = tileStyle;
     final style = switch ((enabled, hovered)) {
       (true, true) => contentStyle.enabledHoveredStyle,
       (true, false) => contentStyle.enabledStyle,
       (false, _) => contentStyle.disabledStyle,
     };
-    final divider = switch (tile) {
-      _ when tile.divider != FTileDivider.none && tile.index != tile.length - 1 => tile.divider,
-      _ when group.divider != FTileDivider.none && tile.index == tile.length - 1 && group.index != group.length - 1 =>
-        group.divider,
+    final divider = switch (focused) {
+      true when tile.index != tile.length - 1 => FTileDivider.full,
+      true when tile.index == tile.length - 1 && group.index != group.length - 1 => FTileDivider.full,
+      false when tile.index != tile.length - 1 => tile.divider,
+      false when tile.index == tile.length - 1 && group.index != group.length - 1 => group.divider,
       _ => FTileDivider.none,
     };
 
@@ -114,7 +115,11 @@ class FTileContent extends StatelessWidget {
           )
         else
           const SizedBox(),
-        if (divider == FTileDivider.none) const SizedBox() else FDivider(style: dividerStyle),
+        switch ((focused, divider)) {
+          (_, FTileDivider.none) => const SizedBox(),
+          (true, _) => FDivider(style: focusedDividerStyle),
+          (false, _) => FDivider(style: dividerStyle),
+        }
       ],
     );
   }