+++ forui/test/src/widgets/radio_golden_test.dart
@@ -27,6 +27,19 @@ void main() {
     });
 
     for (final theme in TestScaffold.themes) {
+      testWidgets('${theme.name} focused', (tester) async {
+        await tester.pumpWidget(
+          TestScaffold(
+            theme: theme.data,
+            child: const FRadio(
+              autofocus: true,
+            ),
+          ),
+        );
+
+        await expectLater(find.byType(TestScaffold), matchesGoldenFile('radio/${theme.name}/focused.png'));
+      });
+
       for (final (enabled, value, error) in [
         (true, true, false),
         (true, true, true),