+++ forui/test/src/widgets/tile/tile_group_golden_test.dart
@@ -195,6 +195,40 @@ void main() {
             );
           });
 
+          testWidgets('focused - ${theme.name} - $divider - $position', (tester) async {
+            await tester.pumpWidget(
+              TestScaffold(
+                theme: theme.data,
+                child: FTileGroup(
+                  label: const Text('Network'),
+                  divider: divider,
+                  children: [
+                    FTile(
+                      autofocus: index == 0,
+                      prefixIcon: FIcon(FAssets.icons.wifi),
+                      title: const Text('WiFi'),
+                      suffixIcon: FIcon(FAssets.icons.chevronRight),
+                      onPress: () {},
+                    ),
+                    FTile(
+                      autofocus: index == 1,
+                      prefixIcon: FIcon(FAssets.icons.bluetooth),
+                      title: const Text('Bluetooth'),
+                      subtitle: const Text('Fee, Fo'),
+                      suffixIcon: FIcon(FAssets.icons.chevronRight),
+                      onPress: () {},
+                    ),
+                  ],
+                ),
+              ),
+            );
+
+            await expectLater(
+              find.byType(TestScaffold),
+              matchesGoldenFile('tile/group/${theme.name}/focused/$divider-$position.png'),
+            );
+          });
+
           testWidgets('disabled - ${theme.name} - $divider - $position', (tester) async {
             await tester.pumpWidget(
               TestScaffold(